Enable SOPS secret decryption on cluster
I don't know how it works yet. The secretRef points to sops-gpg, but this
name is not defined anywhere in redis-auth.yaml file. It was there
initially after running:
kubectl create secret generic sops-gpg
But SOPS CLI changed this file later. The probable is scenario is that
SOPS is encrypting everything, not only the secret value.