Remove redis-auth secret
The secret was created successfully by Flux earlier, but it seems that
the test workflow is not able to validate it. It doesn't produce any
errors, but has a couple of warnings and the last warning's message is:
stdin contains an invalid Secret (redis.redis-auth) - sops: Additional property sops is not allowed

Therefore I'm removing the secret definition for now to check if it will
allow the test workflow to finish successfully.